requirements.txt
feedparser
kafka-python
recipe-scrapers
ingredient-parser-nlp==0.1.0a1
pyspark
snowflake-connector-python
pandas
python-dotenv
spacy

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
